perbaiki bug
Memperbaiki bug dan error serta melakukan test api
pada frontend dan backend

+++ backend/app/api/endpoints/clinic.py
@@ -1,13 +1,15 @@
-from fastapi import APIRouter, Depends
+from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.orm import Session
 from typing import List
 from app.database.session import get_db
 from app.crud import clinic as crud
 from app.schemas import clinic as schemas
+from app.core.security import require_admin, get_current_user
 
 # Kita panggil langsung class-nya agar tidak perlu pakai 'models.'
 from app.models.appointment import Appointment 
 from app.models.clinic import Doctor, Service
+from app.models.patient import Patient
 
 router = APIRouter()
 
@@ -21,7 +23,7 @@ def read_services(db: Session = Depends(get_db)):
 
 # Tambahkan ini di file clinic.py
 @router.post("/doctors", response_model=schemas.DoctorResponse)
-def add_doctor(data: schemas.DoctorBase, db: Session = Depends(get_db)):
+def add_doctor(data: schemas.DoctorBase, db: Session = Depends(get_db), admin: dict = Depends(require_admin)):
     new_doc = Doctor(**data.model_dump())
     db.add(new_doc)
     db.commit()
@@ -29,8 +31,10 @@ def add_doctor(data: schemas.DoctorBase, db: Session = Depends(get_db)):
     return new_doc
 
 @router.delete("/doctors/{doctor_id}")
-def delete_doctor(doctor_id: int, db: Session = Depends(get_db)):
+def delete_doctor(doctor_id: int, db: Session = Depends(get_db), admin: dict = Depends(require_admin)):
     doc = db.query(Doctor).filter(Doctor.id == doctor_id).first()
+    if not doc:
+        raise HTTPException(status_code=404, detail="Dokter tidak ditemukan")
     db.delete(doc)
     db.commit()
     return {"message": "Dokter berhasil dihapus"}
@@ -50,7 +54,7 @@ def get_today_appointments(db: Session = Depends(get_db)):
 
 
 @router.get("/stats/summary")
-def get_admin_stats(db: Session = Depends(get_db)):
+def get_admin_stats(db: Session = Depends(get_db), admin: dict = Depends(require_admin)):
     total_doctors = db.query(Doctor).count()
     total_appointments = db.query(Appointment).count()
     total_patients = db.query(Patient).count()
